requirements.txt
lxml==4.9.1
numpy==1.23.1
pandas==1.4.3
psycopg==3.0.15
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
